refactor: replace fetch with undici and separate concerns
Amazon URLの503エラーを解決するため、fetchをundiciに置き換え。
同時に、ビジネスロジックをsrc/libs/に分離し、責務を明確化。

**変更内容:**
- fetchをundici.requestに変更（Cloud Run環境での503エラー回避を試行）
- HTML escaping追加（XSS対策）
- タイムアウト追加（10秒）
- p-retryパラメータを明示化（minTimeout: 1s, maxTimeout: 4s, factor: 2）
- cache-control fallback修正（'max-age=60, must-revalidate'）
- フォールバック時にURL全体を表示（ホスト名のみから変更）

**リファクタリング:**
- fetchPageMetadata を src/libs/embed/ に分離
- テストも src/libs/embed/ に移動
- src/pages/embed/index.ts はルーティングとHTML生成のみ

**テスト:**
- 5xxエラーでリトライして成功
- 4xxエラーでリトライせずAbort
- 全リトライ失敗でフォールバック
- ネットワークエラーでリトライ
- 無効なURLでもエラーをスローせずフォールバック
